requirements.txt
fastapi
pydantic
pandas
faiss-cpu
numpy
sentence_transformers